+++ tests/core/test_translator.py
@@ -157,7 +157,7 @@ def test_positionalTranslatorIntegration():
     scheme = PositionalScheme()
     saqc = SaQC(data=data, scheme=scheme)
     saqc = saqc.flagMissing(col).flagRange(col, min=3, max=10, flag=DOUBTFUL)
-    flags = saqc.result.flags
+    flags = saqc.flags
 
     for field in flags.columns:
         assert flags[field].astype(str).str.match("^9[012]*$").all()
@@ -177,7 +177,7 @@ def test_dmpTranslatorIntegration():
     scheme = DmpScheme()
     saqc = SaQC(data=data, scheme=scheme)
     saqc = saqc.flagMissing(col).flagRange(col, min=3, max=10)
-    flags = saqc.result.flags
+    flags = saqc.flags
 
     qflags = flags.xs("quality_flag", axis="columns", level=1)
     qfunc = flags.xs("quality_comment", axis="columns", level=1).applymap(
@@ -209,11 +209,11 @@ def test_dmpValidCombinations():
     scheme = DmpScheme()
     saqc = SaQC(data=data, scheme=scheme)
 
-    with pytest.raises(RuntimeError):
-        saqc.flagRange(col, min=3, max=10, cause="SOMETHING_STUPID").result
+    with pytest.raises(ValueError):
+        saqc.flagRange(col, min=3, max=10, cause="SOMETHING_STUPID").flags
 
-    with pytest.raises(RuntimeError):
-        saqc.flagRange(col, min=3, max=10, cause="").result
+    with pytest.raises(ValueError):
+        saqc.flagRange(col, min=3, max=10, cause="").flags
 
 
 def _buildupSaQCObjects():
@@ -241,8 +241,8 @@ def _buildupSaQCObjects():
 def test_translationPreservesFlags():
 
     saqc1, saqc2 = _buildupSaQCObjects()
-    flags1 = saqc1.result.flags_raw
-    flags2 = saqc2.result.flags_raw
+    flags1 = saqc1._flags
+    flags2 = saqc2._flags
 
     for k in flags2.columns:
         got = flags2.history[k].hist
@@ -256,8 +256,8 @@ def test_translationPreservesFlags():
 
 def test_multicallsPreserveHistory():
     saqc1, saqc2 = _buildupSaQCObjects()
-    flags1 = saqc1.result.flags_raw
-    flags2 = saqc2.result.flags_raw
+    flags1 = saqc1._flags
+    flags2 = saqc2._flags
 
     # check, that the `History` is duplicated
     for col in flags2.columns:
@@ -280,8 +280,8 @@ def test_positionalMulitcallsPreserveState():
     saqc1, saqc2 = _buildupSaQCObjects()
 
     scheme = PositionalScheme()
-    flags1 = saqc1.result.flags_raw
-    flags2 = saqc2.result.flags_raw
+    flags1 = saqc1._flags
+    flags2 = saqc2._flags
     tflags1 = scheme.backward(flags1).astype(str)
     tflags2 = scheme.backward(flags2).astype(str)